# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal EOD Market API",
    "description": "收盘数据服务；CN 交易日北京时间 15:10、17:10、19:10 三轮采集；美股行业 ETF 代理排名和可编辑 AI 观察池。",
    "version": "0.4.1"
  },
  "paths": {
...
    "/v1/catalog-coverage": {
      "get": {
        "summary": "Catalog Coverage",
        "operationId": "catalog_coverage_v1_catalog_coverage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/deployment-status": {
      "get": {
        "summary": "Deployment Status",
        "operationId": "deployment_status_v1_deployment_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}